% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{_chisq_test}
\alias{_chisq_test}
\alias{chisq_test}
\title{Compute the Chi-Squared statistic for a given column, grouped by another column.}
\usage{
chisq_test(values, groups)
}
\arguments{
\item{\code{values}}{The vector of values}

\item{\code{groups}}{The vector of groups, corresponding to the values}
}
\value{
A numeric vector of the statistic values for each group, named by group.
}
\description{
Computes (1 - p-value) * 100 for each group, using a
Chi-Squared statistic to obtain the p-value.
}
