% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{counts}
\alias{counts}
\title{Compute the population and group counts for a vector of values and
corresponding vector of group identifiers.}
\arguments{
\item{\code{values}}{The vector of values}

\item{\code{groups}}{The vector of groups, corresponding to the values}
}
\value{
A std::pair where the first item is the map of value -> count for the population,
and the second item is the map of group -> value -> count for the groups.
}
\description{
Compute the population and group counts for a vector of values and
corresponding vector of group identifiers.
}
