% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_config.R
\name{def_composite_score}
\alias{def_composite_score}
\alias{def_composite_score_q}
\title{Define a composite score.}
\usage{
def_composite_score(...)

def_composite_score_q(name, compute, ...)
}
\description{
Define a composite score.
}
