% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_config.R
\name{def_group_metric}
\alias{def_group_metric}
\alias{def_group_metric_q}
\title{Define a group metric.}
\usage{
def_group_metric(name, compute, compute.group, ...)

def_group_metric_q(name, compute, compute.group, ...)
}
\description{
A group metric is a special metric that does not produce entity and 
group scores in the ordinary way. Instead, the entity metric values are
combined at the group level by a function \code{compute.group} of three
arguments: the data table, the metric column name, and the group column name.
}
