% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_config.R
\name{def_metric}
\alias{def_metric}
\alias{def_metric_q}
\title{Define a metric.}
\usage{
def_metric(..., control_for = NULL)

def_metric_q(name, compute, control_for = NULL)
}
\description{
Define a metric.
}
