% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_config.R
\name{def_parameters}
\alias{def_parameters}
\alias{def_parameter_q}
\title{Define global model parameters.}
\usage{
def_parameters(...)

def_parameter_q(name, value)
}
\description{
Supported parameters:
\itemize{
  \item \code{entity_name}: The column containing the entity ID (required).
  \item \code{group_name}: The column containing the group ID (required).
  \item \code{cap_entity_score}: Entity scores above this number 
   (in absolute value) will be capped (default = \code{Inf}).
  \item \code{min_group_size}: Groups with size smaller than this number
    will be discarded (default = \code{Inf}).
}
}
