% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{_entropy_disc}
\alias{_entropy_disc}
\alias{entropy_disc}
\title{Compute information entropy of the values in each group.}
\usage{
entropy_disc(values, groups, type)
}
\arguments{
\item{\code{values}}{The vector of values}

\item{\code{groups}}{The vector of groups, corresponding to the values}

\item{\code{type}}{The type of ratio to compute: "normal" or "frequent"}
}
\value{
A numeric vector of the values for each group, named by group.
Entropy is maximal when all events are equaly probable.
Normalized entropy is also known as Efficiency. It is the computed entropy
   divided by the maximum possible entropy for N events = log(N)
See, eg, https://en.wikipedia.org/wiki/Entropy_(information_theory)
}
\description{
Compute information entropy of the values in each group.
}
