% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{group_size}
\alias{group_size}
\title{Compute the size of a group from its counts.}
\arguments{
\item{\code{group}}{A map from the counts of the group}
}
\value{
The size of the group
}
\description{
Compute the size of a group from its counts.
}
