% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_util.R
\name{ks.stat}
\alias{ks.stat}
\title{Kolmogorov-Smirnov statistic}
\usage{
ks.stat(x, y)
}
\description{
Computes the Kolmogorov-Smirnov statistic for numerical data.
}
\details{
Unlike \code{ks.test}, this function supports both continuous and 
discrete numerical data. However, since the p-values are not reliable in the
latter case, no p-values are computed.
}
