% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_env.R
\name{lazy_env}
\alias{lazy_env}
\alias{is_lazy_env}
\title{Lazy environment}
\usage{
lazy_env(promises)

is_lazy_env(x)
}
\value{
An object of class "lazy_env"
}
\description{
An environment that loads its objects lazily. Its main purpose is to create
wrappers around files or databases too large to fit comfortably in memory.
}
\seealso{
\code{lazy_fread}, \code{lazy_sql}
}
