% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_env_backends.R
\name{lazy_sql}
\alias{lazy_sql}
\title{Lazy environment from SQL database}
\usage{
lazy_sql(conn, table_name, order_by = NULL)
}
\arguments{
\item{\code{conn}}{DBI connection}

\item{\code{table_name}}{table to read from}
}
\description{
The objects in the environment are the columns of a table in the database.
}
