% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_util.R
\name{Mean}
\alias{Mean}
\alias{Sd}
\alias{Scale}
\alias{Quantile}
\title{Missing value variants}
\usage{
Mean(x, ...)

Sd(x)

Scale(x)

Quantile(x, ...)
}
\description{
These functions are variants of the standard R functions that ignore
missing values.
}
