% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{_uniq_cont}
\alias{_uniq_cont}
\alias{uniq_cont}
\title{Computes the ratio of the coefficient of variation for each group to the
population.}
\usage{
uniq_cont(values, groups)
}
\arguments{
\item{\code{values}}{The vector of values}

\item{\code{groups}}{The vector of groups, corresponding to the values}
}
\value{
A numeric vector of the ratio values for each group, named by group.
}
\description{
Computes the ratio of the coefficient of variation for each group to the
population.
}
