% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{write_sql_db}
\alias{write_sql_db}
\title{Write multiple tables to a SQL database}
\usage{
write_sql_db(conn, tables)
}
\arguments{
\item{\code{conn}}{DBI connection}

\item{\code{tables}}{list of data tables, each of one of the following types:
\itemize{
  \item data frame (including data.table and ffdf)
  \item environment
}}
}
\description{
Write multiple tables to a SQL database
}
\details{
Warning: if any of the tables already exist, they will be dropped!
}
\seealso{
\code{read_sql_db}
}
