% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_util.R
\name{z_score}
\alias{z_score}
\title{Z score}
\usage{
z_score(x, cap = NULL)
}
\description{
Unlike \code{scale}, this function ignores missing values. Also, the scores
may be capped at some threshold value.
}
